# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: nextStop Candidate Search API
  version: 1.1.0-draft
  description: |
    Draft architecture-gate contract. The v1 wire shape is unchanged across two
    candidate modes. With foodChain omitted or null, the backend emits one bounded
    charging campus candidate per stable campus ID. With foodChain non-null, it emits
    complete-link fine-park candidates after the exact selected-chain proximity
    predicate from a pinned OSM projection. The backend performs exact route
    corridor filtering. MapKit on iOS calculates actual driving distance, applies
    the exact selected distance range, and performs final distance-only ranking.
    Supported physical devices obtain a 15-minute search token through Apple App
    Attest. Challenges are single use and expire after three minutes. A Debug
    Simulator obtains the same short-lived token shape from a loopback-only Mac
    broker; that local broker is not part of this public HTTPS API. Behind the
    public origin, Nginx routes only `/v1/auth/app-attest/*` to the isolated
    authentication service on VM loopback port 3001. Health and candidate search
    remain on the read-only search API on VM loopback port 3000.
    For iOS 27 proofs that carry Apple's validation extensions, validation
    category and bundle version must appear together and are checked on every
    attestation and assertion. Development keys accept category 3; production
...
servers:
  - url: https://api.nextstop.tech
paths:
...
  /v1/charging-parks/search:
    post:
      operationId: searchChargingParkCandidates
      summary: Return a stable page of route-corridor charging candidates
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Stable candidate page; this is not the final ranked list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing, malformed, expired, or invalid search bearer credential.
          headers:
            WWW-Authenticate:
              schema:
                type: string
                const: 'Bearer realm="nextstop-search"'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '413':
          description: Request body exceeds 512 KiB.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Cursor does not belong to the supplied snapshot/request.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Rate limited or all bounded search slots are occupied.
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 1
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '503':
          description: No valid charging or requested food-POI projection is available.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A server-signed search access token issued after App Attest verification
        and valid for exactly 15 minutes. A Debug Simulator may present the same
        token shape after its loopback Mac broker mints one through Google Cloud
        IAP. During the private staging migration only, the server may also accept
        the former revocable bearer when `ALLOW_LEGACY_STAGING_BEARER=true`; that
        compatibility credential is not part of distributed App Attest builds.
  responses:
    BadRequest:
      description: Invalid route or criteria.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    SearchRequest:
      type: object
      additionalProperties: false
      required: [requestId, route, criteria]
      properties:
        requestId:
          type: string
          format: uuid
          description: Fresh per-operation identifier; never a device/user ID.
        route:
          $ref: '#/components/schemas/RouteLineString'
        criteria:
          $ref: '#/components/schemas/SearchCriteria'
        page:
          type: object
          additionalProperties: false
          properties:
            snapshotToken:
              type: string
              maxLength: 512
            cursor:
              type: string
              maxLength: 512
          description: Omit both fields on the first page; echo server values later.
    RouteLineString:
      type: object
      additionalProperties: false
      required: [type, coordinates]
      properties:
        type:
          const: LineString
        coordinates:
          type: array
          minItems: 2
          maxItems: 8000
          description: >
            Detailed MapKit route. Every coordinate must be within the supported
            European envelope; route length is limited to 2,500 km and individual
            segments to 250 km. Geometry is rejected, never silently simplified.
          items:
            type: array
            prefixItems:
              - type: number
                minimum: -25
                maximum: 45
                description: Longitude.
              - type: number
                minimum: 34
                maximum: 72
                description: Latitude.
            minItems: 2
            maxItems: 2
    SearchCriteria:
      type: object
      additionalProperties: false
      required:
        - distanceRangeMeters
        - minimumChargingPoints
        - minimumPowerKW
      properties:
        distanceRangeMeters:
          oneOf:
            - $ref: '#/components/schemas/Range15To50'
            - $ref: '#/components/schemas/Range50To100'
            - $ref: '#/components/schemas/Range100To150'
          description: Used for safe batching; exact filtering occurs on iOS.
        minimumChargingPoints:
          type: integer
          enum: [2, 4, 6, 8, 10, 12, 16, 20]
          description: >
            Applied after EVSEs below minimumPowerKW are removed and the remainder
            is deduplicated across the selected candidate unit: the whole campus
            when foodChain is omitted or null, otherwise one fine park.
        minimumPowerKW:
          type: integer
          enum: [11, 22, 50, 100, 150, 200, 250, 300, 350, 400]
          description: >
            Applied to each EVSE before candidate-wide deduplication, operator
            counts, and minimumChargingPoints. Availability remains informational.
        foodChain:
          type: [string, 'null']
          enum: [null, mcdonalds, burger_king, kfc, subway]
          description: >
            Omitted or null selects bounded no-food campus candidates. A non-null
            value selects complete-link fine-park candidates and requires a matching
            OSM restaurant within 500 m of the power-filtered fine-park navigation
            coordinate.
    Range15To50:
      type: object
      additionalProperties: false
      required: [minimum, maximum]
      properties:
        minimum: { const: 15000 }
        maximum: { const: 50000 }
    Range50To100:
      type: object
      additionalProperties: false
      required: [minimum, maximum]
      properties:
        minimum: { const: 50000 }
        maximum: { const: 100000 }
    Range100To150:
      type: object
      additionalProperties: false
      required: [minimum, maximum]
      properties:
        minimum: { const: 100000 }
        maximum: { const: 150000 }
    SearchResponse:
      type: object
      additionalProperties: false
      required: [snapshotToken, candidates, coverage, generatedAt, attributions]
      properties:
        snapshotToken:
          type: string
        nextCursor:
          type: [string, 'null']
        generatedAt:
          type: string
          format: date-time
        candidates:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/ChargingParkCandidate'
        coverage:
          $ref: '#/components/schemas/Coverage'
        attributions:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/DataAttribution'
    ChargingParkCandidate:
      type: object
      additionalProperties: false
      description: |
        Compatibility name for the v1 candidate shape. With foodChain omitted or
        null this is one ChargingCampus (fine-park seeds joined by <=200 m
        cross-edges while the union diameter remains <=500 m). With foodChain
        non-null this is one
        complete-link ChargingPark with maximum diameter <=200 m and an exact food
        match. The backend chooses this identity before charging filters and paging.
        coordinate is the power-filtered display coordinate; navigationCoordinate
        is the qualifying member/access point used consistently for corridor,
        origin lower bound, MapKit distance, and handoff. chargingPoints, operator
        counts, availability, maximumPowerKW, and locationLookups all cover that
        selected candidate unit after power filtering and EVSE deduplication.
      required:
        - id
        - name
        - coordinate
        - navigationCoordinate
        - distanceFromRouteMeters
        - straightLineLowerBoundMeters
        - chargingPoints
        - availability
        - maximumPowerKW
        - operators
        - operatorChargingPoints
        - locationLookups
        - sources
        - dataUpdatedAt
      properties:
        id:
          type: string
          format: uuid
          description: >
            Stable charging-campus-v1 ID when foodChain is omitted or null;
            otherwise the stable fine-park ID.
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: >
            Power-filtered candidate name; campus-wide when foodChain is omitted or
            null.
        coordinate: { $ref: '#/components/schemas/Coordinate' }
        navigationCoordinate: { $ref: '#/components/schemas/Coordinate' }
        distanceFromRouteMeters:
          type: integer
          minimum: 0
          maximum: 5000
          description: >
            Exact geodesic distance from navigationCoordinate to the supplied route
            line.
        preliminaryRouteProgressMeters:
          type: integer
          minimum: 0
          description: Never display or treat as actual driving distance.
        straightLineLowerBoundMeters:
          type: integer
          minimum: 0
          description: >
            Geodesic origin-to-navigationCoordinate lower bound; never display as
            driving distance.
        chargingPoints:
          type: integer
          minimum: 1
          description: >
            Deduplicated EVSEs meeting minimumPowerKW across the whole campus when
            foodChain is omitted or null, otherwise across this fine park.
        availability: { $ref: '#/components/schemas/ParkAvailability' }
        maximumPowerKW:
          type: integer
          minimum: 1
          description: >
            Maximum power among qualifying EVSEs across the selected candidate unit.
        operators:
          type: array
          uniqueItems: true
          items: { type: string }
        operatorChargingPoints:
          type: array
          minItems: 1
          description: >
            Exact-name operator partitions of candidate-wide deduplicated EVSEs
            meeting minimumPowerKW; the sum equals chargingPoints. Campus mode must
            not sum pre-aggregated fine-park counts.
          items:
            $ref: '#/components/schemas/OperatorChargingPoints'
        locationLookups:
          type: array
          minItems: 1
          description: >
            Power-filtered authority locations across the whole selected candidate
            unit, used only for on-device Apple POI matching.
          items:
            $ref: '#/components/schemas/ChargingLocationLookup'
        sources:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SourceSummary' }
        dataUpdatedAt: { type: string, format: date-time }
        foodPOI:
          description: >
            Absent or null in campus mode; the exact <=500 m OSM match for this
            fine park in food mode.
          oneOf:
            - $ref: '#/components/schemas/FoodPOI'
            - type: 'null'
    ChargingLocationLookup:
      type: object
      additionalProperties: false
      required: [id, operatorName, coordinate, address]
      properties:
        id: { type: string, format: uuid }
        operatorName: { type: string, minLength: 1, maxLength: 200 }
        coordinate: { $ref: '#/components/schemas/Coordinate' }
        address:
          type: object
          additionalProperties: false
          properties:
            street: { type: string, minLength: 1, maxLength: 200 }
            houseNumber: { type: string, minLength: 1, maxLength: 50 }
            postalCode: { type: string, minLength: 1, maxLength: 20 }
            city: { type: string, minLength: 1, maxLength: 200 }
    FoodPOI:
      type: object
      additionalProperties: false
      required: [id, chain, name, coordinate, distanceFromChargingParkMeters, sourceRecordURL]
      properties:
        id: { type: string, minLength: 1 }
        chain:
          type: string
          enum: [mcdonalds, burger_king, kfc, subway]
        name: { type: string, minLength: 1, maxLength: 200 }
        coordinate: { $ref: '#/components/schemas/Coordinate' }
        distanceFromChargingParkMeters:
          type: integer
          minimum: 0
          maximum: 500
          description: >
            Exact geodesic distance to the power-filtered fine-park navigation
            coordinate.
        openingHours: { type: [string, 'null'] }
        sourceRecordURL: { type: string, format: uri }
    DataAttribution:
      type: object
      additionalProperties: false
      required: [id, name, notice, licenseName, licenseURL]
      properties:
        id: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        notice: { type: string, minLength: 1 }
        licenseName: { type: string, minLength: 1 }
        licenseURL: { type: string, format: uri }
        transportName: { type: [string, 'null'] }
        transportURL: { type: [string, 'null'], format: uri }
    OperatorChargingPoints:
      type: object
      additionalProperties: false
      required: [name, chargingPoints]
      properties:
        name: { type: string, minLength: 1 }
        chargingPoints: { type: integer, minimum: 1 }
    Coordinate:
      type: object
      additionalProperties: false
      required: [latitude, longitude]
      properties:
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
    ParkAvailability:
      type: object
      additionalProperties: false
      required:
        - knownAvailable
        - knownUnavailable
        - unknown
        - total
        - complete
      properties:
        knownAvailable: { type: integer, minimum: 0 }
        knownUnavailable: { type: integer, minimum: 0 }
        unknown: { type: integer, minimum: 0 }
        total: { type: integer, minimum: 1 }
        complete: { type: boolean }
        observedAt: { type: [string, 'null'], format: date-time }
    SourceSummary:
      type: object
      additionalProperties: false
      required: [id, name, qualityTier, staticObservedAt]
      properties:
        id: { type: string }
        name: { type: string }
        qualityTier:
          type: string
          enum: [operator, authority, open_data, community]
        staticObservedAt: { type: string, format: date-time }
        liveObservedAt: { type: [string, 'null'], format: date-time }
    Coverage:
      type: object
      additionalProperties: false
      required: [status, activeSources, unavailableSources, projectionUpdatedAt]
      properties:
        status:
          type: string
          enum: [complete, degraded, stale]
        activeSources:
          type: array
          items: { type: string }
        unavailableSources:
          type: array
          items: { type: string }
        projectionUpdatedAt: { type: string, format: date-time }
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, errorId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        errorId:
          type: string
          minLength: 16
          maxLength: 128
          pattern: '^[A-Za-z0-9_-]+$'
          description: Opaque support identifier; do not infer UUID semantics.